Implementing router to add a new page - Adding react-router-dom library to allow create routes - Adjusting webpack dev server config to allow reloads on new routes - Changing App.js component 	- removing all home page logic that will be implemented in a 	page component 	- adding router library components to create home page route, 	and movie details route - Creating Home page component 	- all home page logic moved to this component (loading, state 	management and presenting page components) - Layout changes 	- creating button component to reuse in different components 	- addin a see details button in each movie summary 	- adding property to load more button to define its style
