+++ package.json
@@ -40,6 +40,7 @@
     "nodemon": "^1.17.3",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.1.2",
     "redis": "^2.8.0",
     "style-loader": "^0.23.0",
     "styled-components": "^5.0.0",
