+++ src/client/components/Button/index.js
@@ -1,0 +1,3 @@
+import Button from "./Button";
+
+export default Button;
