+++ src/client/components/LoadMore/LoadMore.jsx
@@ -1,29 +1,19 @@
 import React from "react";
 import styled from "styled-components";
+import Button from "../Button";
 
 const Container = styled.div`
   display: flex;
   justify-content: center;
   padding: 10px;
 `;
-const LoadMoreButton = styled.button`
-  padding: 10px 25px;
-  background-color: #eee;
-  border: 1px solid;
-  border-color: #d6d6d6;
-  border-radius: 2px;
-  font-size: 18px;
-  cursor: pointer;
-  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.15);
-  &:hover {
-    box-shadow: 2px 4px 8px 2px rgba(0, 0, 0, 0.2);
-  }
-`;
 
 const LoadMore = ({ loadMore }) => {
   return (
     <Container>
-      <LoadMoreButton onClick={loadMore}>See more</LoadMoreButton>
+      <Button large onClick={loadMore}>
+        See more
+      </Button>
     </Container>
   );
 };
